requirements.txt
PyYAML~=6.0.2
deepdiff~=8.0.1
tabulate~=0.9.0
openshift